+++ src/layout/fallback_layout.h
@@ -52,12 +52,6 @@ protected:
     void addVideo(zmm::Ref<CdsObject> obj, zmm::String rootpath);
     void addImage(zmm::Ref<CdsObject> obj, zmm::String rootpath);
     void addAudio(zmm::Ref<CdsObject> obj);
-#ifdef HAVE_LIBDVDNAV
-    zmm::Ref<CdsObject> prepareChapter(zmm::Ref<CdsObject> obj, int title_idx,
-                                       int chapter_idx);
-    void addDVD(zmm::Ref<CdsObject> obj);
-    zmm::String mpeg_mimetype;
-#endif
 #ifdef YOUTUBE
     void addYouTube(zmm::Ref<CdsObject> obj);
 #endif