+++ src/metadata/fanart_handler.h
@@ -2,7 +2,7 @@
     
     MediaTomb - http://www.mediatomb.cc/
     
-    dvd_handler.h - this file is part of MediaTomb.
+    fanart_handler.h - this file is part of MediaTomb.
     
     Copyright (C) 2005 Gena Batyan <bgeradz@mediatomb.cc>,
                        Sergey 'Jin' Bostandzhyan <jin@mediatomb.cc>