+++ src/metadata_handler.h
@@ -60,7 +60,6 @@
 #define CONTENT_TYPE_PCM        "pcm"
 #define CONTENT_TYPE_AVI        "avi"
 #define CONTENT_TYPE_MPEG       "mpeg"
-#define CONTENT_TYPE_DVD        "dvd"
 #define CONTENT_TYPE_QUICKTIME  "quicktime"
 #define CONTENT_TYPE_MKV        "mkv"
 #define CONTENT_TYPE_MKA        "mka"