+++ src/scripting/script.h
@@ -44,8 +44,7 @@
 typedef enum
 {
     S_IMPORT = 0,
-    S_PLAYLIST,
-    S_DVD
+    S_PLAYLIST
 } script_class_t;
 
 typedef enum